+++ components/common/TokenGateForm/index.ts
@@ -1,0 +1,3 @@
+import TokenGateForm from './TokenGateForm';
+
+export default TokenGateForm;
